+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added new validators in `HeatSimulation`: no structures with dimensions of zero size, no all-Neumann boundary conditions, non-empty simulation domain.
 
+- A batch of Mode Solvers can be run concurrently using `tidy3d.plugins.mode.web.run_batch()`.
+
 ### Changed
 - Revert forbidden `"` in component names.